Update dct2Test3.cpp
Update dct2Test3.cpp

+++ test/dct2Test3.cpp
@@ -39,7 +39,8 @@ constexpr static auto get_offset( ExPo execution_policy,
         std::vector<TinyDIP::Image<ElementT>> outputs;
         outputs.resize(dictionary_y.size());
         #pragma omp parallel for
-        for (size_t i = 0; i < dictionary_y.size(); ++i)
+        //  C3016 'i': index variable in OpenMP 'for' statement must have signed integral type
+        for (int i = 0; i < dictionary_y.size(); ++i)
         {
             outputs[i] = dictionary_y[i] * weights[i];
         }
